Fixes part 2
