+++ client/Packages/com.beamable.server/Editor/UI/MicroserviceWindow.cs
@@ -104,9 +104,8 @@ namespace Beamable.Editor.Microservice.UI
 			_windowRoot = uiAsset.CloneTree();
 			_windowRoot.AddStyleSheet($"{Directories.BEAMABLE_SERVER_PACKAGE_EDITOR_UI}/MicroserviceWindow.uss");
 			_windowRoot.name = nameof(_windowRoot);
-#if UNITY_2019_1_OR_NEWER
-			_windowRoot.TryAddScrollViewAsMainElement(ScrollViewMode.Vertical);
-#endif
+			_windowRoot.TryAddScrollViewAsMainElement();
+
 			root.Add(_windowRoot);
 
 			bool localServicesAvailable = Model?.AllLocalServices != null;
