+++ client/Packages/com.beamable/Editor/UI/Toolbox/ToolboxWindow.cs
@@ -111,9 +111,7 @@ namespace Beamable.Editor.Toolbox.UI
 			_windowRoot = uiAsset.CloneTree();
 			_windowRoot.AddStyleSheet($"{BASE_PATH}/ToolboxWindow.uss");
 			_windowRoot.name = nameof(_windowRoot);
-#if UNITY_2019_1_OR_NEWER
-			_windowRoot.TryAddScrollViewAsMainElement(ScrollViewMode.Vertical);
-#endif
+			_windowRoot.TryAddScrollViewAsMainElement();
 
 			root.Add(_windowRoot);
